refactor: use shared function to initialize models
