+++ backend/src/core/models/index.ts
@@ -10,3 +10,4 @@ export * from './statistic'
 export * from './protected_message'
 export * from './unsubscriber'
 export * from './agency'
+export * from './initialize-models'
