+++ backend/src/core/models/initialize-models.ts
@@ -1,0 +1,70 @@
+import { Sequelize } from 'sequelize-typescript'
+import {
+  Credential,
+  JobQueue,
+  Campaign,
+  Worker,
+  User,
+  UserFeature,
+  UserCredential,
+  UserDemo,
+  Statistic,
+  ProtectedMessage,
+  Unsubscriber,
+  Agency,
+} from '@core/models'
+import {
+  EmailMessage,
+  EmailTemplate,
+  EmailOp,
+  EmailBlacklist,
+  EmailFromAddress,
+} from '@email/models'
+import { SmsMessage, SmsTemplate, SmsOp } from '@sms/models'
+import {
+  BotSubscriber,
+  TelegramMessage,
+  TelegramOp,
+  TelegramSubscriber,
+  TelegramTemplate,
+} from '@telegram/models'
+
+export const initializeModels = (sequelize: Sequelize): void => {
+  const coreModels = [
+    Credential,
+    JobQueue,
+    Campaign,
+    Worker,
+    User,
+    UserFeature,
+    UserCredential,
+    UserDemo,
+    Statistic,
+    Unsubscriber,
+    Agency,
+  ]
+  const emailModels = [
+    EmailMessage,
+    EmailTemplate,
+    EmailOp,
+    EmailBlacklist,
+    ProtectedMessage,
+    EmailFromAddress,
+  ]
+  const smsModels = [SmsMessage, SmsTemplate, SmsOp]
+  const telegramModels = [
+    BotSubscriber,
+    TelegramOp,
+    TelegramMessage,
+    TelegramTemplate,
+    TelegramSubscriber,
+  ]
+  sequelize.addModels([
+    ...coreModels,
+    ...emailModels,
+    ...smsModels,
+    ...telegramModels,
+  ])
+}
+
+export default initializeModels
