+++ backend/src/test-utils/sequelize-loader.ts
@@ -1,33 +1,6 @@
 import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
 import config from '@core/config'
-import {
-  Credential,
-  JobQueue,
-  Campaign,
-  Worker,
-  User,
-  UserFeature,
-  UserCredential,
-  UserDemo,
-  Statistic,
-  ProtectedMessage,
-  Unsubscriber,
-} from '@core/models'
-import {
-  EmailMessage,
-  EmailTemplate,
-  EmailOp,
-  EmailBlacklist,
-  EmailFromAddress,
-} from '@email/models'
-import { SmsMessage, SmsTemplate, SmsOp } from '@sms/models'
-import {
-  BotSubscriber,
-  TelegramMessage,
-  TelegramOp,
-  TelegramSubscriber,
-  TelegramTemplate,
-} from '@telegram/models'
+import { Credential, initializeModels } from '@core/models'
 
 import { DefaultCredentialName } from '@core/constants'
 import { formatDefaultCredentialName } from '@core/utils'
@@ -41,40 +14,7 @@ const sequelizeLoader = async (dbName: string): Promise<Sequelize> => {
     pool: config.get('database.poolOptions'),
   } as SequelizeOptions)
 
-  const coreModels = [
-    Credential,
-    JobQueue,
-    Campaign,
-    Worker,
-    User,
-    UserFeature,
-    UserCredential,
-    UserDemo,
-    Statistic,
-    Unsubscriber,
-  ]
-  const emailModels = [
-    EmailMessage,
-    EmailTemplate,
-    EmailOp,
-    EmailBlacklist,
-    ProtectedMessage,
-    EmailFromAddress,
-  ]
-  const smsModels = [SmsMessage, SmsTemplate, SmsOp]
-  const telegramModels = [
-    BotSubscriber,
-    TelegramOp,
-    TelegramMessage,
-    TelegramTemplate,
-    TelegramSubscriber,
-  ]
-  sequelize.addModels([
-    ...coreModels,
-    ...emailModels,
-    ...smsModels,
-    ...telegramModels,
-  ])
+  initializeModels(sequelize)
 
   try {
     await sequelize.sync()
